+++ lost_ds/segmentation/semantic_seg.py
@@ -45,7 +45,7 @@ def semantic_segmentation(order, dst_dir, fill_value, df, anno_dtypes=['polygon'
                           use_empty=False, lbl_col='anno_lbl', 
                           dst_path_col='seg_path', dst_lbl_col='seg_lbl', 
                           line_thickness=None, radius=None, filesystem=None, 
-                          parallel=-1):
+                          numeric_filename=False, parallel=-1):
     '''Create semantic segmentations from polygon-annos
     
     Args:
@@ -70,6 +70,9 @@ def semantic_segmentation(order, dst_dir, fill_value, df, anno_dtypes=['polygon'
             of dtype point. Only takes effect when anno_dtypes contains 'point'
         filesystem (fsspec.filesystem, FileMan): filesystem to use. Use local
             if not initialized
+        numeric_filename (bool): use same filename for seg and img if False. If
+            True a numeric 12-digit filename will be generated. Usefull if some 
+            files have same filenames.
         
     Returns:
         pd.DataFrame: The original dataframe with new column (dst_col) 
@@ -91,7 +94,7 @@ def semantic_segmentation(order, dst_dir, fill_value, df, anno_dtypes=['polygon'
     # keep empty images if flag is set
     df = pd.concat([df, df_empty])
     
-    def generate_seg(image_path, img_df):
+    def generate_seg(i, image_path, img_df):
         geom = LOSTGeometries()
         if not fs.exists(image_path):
             raise Exception('Image {} does not exist'.format(image_path))
@@ -106,8 +109,10 @@ def semantic_segmentation(order, dst_dir, fill_value, df, anno_dtypes=['polygon'
                                                  row.anno_format,row.anno_style,
                                                  line_thickness=line_thickness, 
                                                  radius=radius)
-        
-        filename = image_path.split('/')[-1].split('.')[0] + '.png'
+        if numeric_filename:
+            filename = f'{i:012d}.png'
+        else:
+            filename = image_path.split('/')[-1].split('.')[0] + '.png'
         seg_path = os.path.join(dst_dir, filename)
         fs.write_img(segmentation, seg_path)
         return image_path, seg_path
@@ -115,14 +120,14 @@ def semantic_segmentation(order, dst_dir, fill_value, df, anno_dtypes=['polygon'
     fs.makedirs(dst_dir, exist_ok=True)
     
     if parallel:
-        paths = Parallel(n_jobs=parallel)(delayed(generate_seg)(path, img_df) 
-                                        for path, img_df in tqdm(
-                                            df.groupby('img_path'), 
+        paths = Parallel(n_jobs=parallel)(delayed(generate_seg)(i, path, img_df) 
+                                        for i, (path, img_df) in tqdm(
+                                            enumerate(df.groupby('img_path')), 
                                             desc='segmentation'))
         path_map = {im_path: seg_path for im_path, seg_path in paths}
     else:
         path_map = dict()
-        for path, img_df in tqdm(df.groupby('img_path'), desc='segmentation'):
+        for i, (path, img_df) in tqdm(enumerate(df.groupby('img_path')), desc='segmentation'):
             p1, p2 = generate_seg(path, img_df)
             path_map[p1] = p2